Moved POST request code to HTTPPost
